requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-chartjs==2.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-jazzmin==2.6.0
django-material==1.11.3
django-material-admin==1.8.5
django-mathfilters==1.0.0
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.1
Pillow==9.4.0
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.14
